Convert Input Count to Frame Count
When recording or playing a movie file, the status bar would display the data relating to how many times the process had polled for inputs. This change simply converts this data into a frame count, as that information is far more useful and easily understood when creating a TAS. I have thoroughly tested to make sure this count does not repeat or skip frames.