Convert Input Count to Frame Count
When recording or playing a movie file, the status bar displays the "input count" which is the number of times the process polls for button and circle pad inputs. By multiplying this data by a coefficient, it can be converted into a frame count which is far more useful and easily understandable when creating a TAS. I have thoroughly tested this value with a number of games for an extended period of time and found no instances where a frame count was skipped or repeated.